requirements.txt
Django==5.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dj-database-url==0.5.0
psycopg2==2.9.9
gunicorn==23.0.0
